Tuple ID tests
Added tuple ID tests using the built in test driver.  The driver is terrible, but we don't have JUnit integration, and I don't think we can add it in this project.  Tests are mostly for sanity sake, feel free to see how they are integrated int he test make file.