qNLJ Untested Implementation
Implementation is theoretically complete, but is completely untested.
To test this code, the changes from issue 16 are required, meaning the
pull request #20 must be completed.

This is because the code requires access to multiple tables from a
single database to perform as intended, and the master branch from which
this branch is based off can only insert into a single hard-coded table
per database.

See Issue #18